(SERVER-2265) Refactor: move import
