+++ lib/puppetserver/ca/cli.rb
@@ -3,7 +3,7 @@ require 'puppetserver/ca/version'
 require 'puppetserver/ca/logger'
 require 'puppetserver/ca/action/clean'
 require 'puppetserver/ca/action/create'
-require 'puppetserver/ca/import_action'
+require 'puppetserver/ca/action/import'
 require 'puppetserver/ca/action/generate'
 require 'puppetserver/ca/revoke_action'
 require 'puppetserver/ca/list_action'
