+++ spec/puppetserver/ca/generate_action_spec.rb
@@ -4,7 +4,7 @@ require 'utils/ssl'
 require 'tmpdir'
 require 'fileutils'
 
-require 'puppetserver/ca/import_action'
+require 'puppetserver/ca/action/import'
 require 'puppetserver/ca/cli'
 
 RSpec.describe Puppetserver::Ca::Action::Generate do
